Use tree structure for topic tree for randomized sibling updates
We now use a tree for the data cache. This allows us, when doing
randomized updates to Diffusion, to choose data from topics that are
the sibling of the currently (semi-randomly chosen) topic.

Randomization isn't truly random. Every time through a loop we shuffle
all topics into a random order and update each one in turn.